Added better support for other mods to add own entity synchronization.
This means, that you can now register an IEntitySynchronizer. When the
player is rendered, your IEntitySynchronizer instance first checks if it
should even be called to synchronize entities, with the method
IEntitySynchronizer#appliesToMorph. Then, if this method returns true,
IEntitySynchronizer#applyToMorph is being called. Here you can set
various attributes, e. g. you can set the inventory of a zombie so that
it is synchronized with the player.