Fixed "skeleton" hand item bug.
This is a bug that caused the issue of foxes, villagers, etc. not having
items in their mouth (provided the offhand was empty), as the hands were
switched because skeletons have their hands switched. This bug is now
fixed, so things should appear in the mouth of a fox now if you hold an
item in your hand.