Removed code comment.
For anyone wondering why this was here in the first place, it was a
direct citation of a code comment included in this video:
https://www.youtube.com/watch?v=k238XpMMn38&t=103s

I should note that this is of course a hyperbole, but this does not in
any way change the fact that this is extremely insulting to a bunch of
people. As such, this code comment is not acceptable in this codebase,
and it shall henceforth be removed. It doesn't reflect my personal
opinions either, it is highly unprofessional, and it directly conflicts
with the concept of the categorical imperative.

This removal is long overdue.
I apologize for this code comment.

- Budschie